pipeline/requirements.txt
pyyaml>=6.0.1
jira>=3.5.1
pytest>=7.4.0
typing-extensions>=4.7.1